Fix mapping memory at specific address on FreeBSD
FreeBSD allows to specify the MAP_EXCL flag with MAP_FIXED, so that the
mapping will fail if any part of the region is already mapped. This is
consistent with the behaviour expected by OpenJDK.

This work was sponsored by: The FreeBSD Foundation

Back-ported additional changes to anon_mmap as well.